rustigram_api::methods::chat_management

Struct PromoteChatMemberParams

pub struct PromoteChatMemberParams {

    pub is_anonymous: Option<bool>,
    pub can_manage_chat: Option<bool>,
    pub can_delete_messages: Option<bool>,
    pub can_manage_video_chats: Option<bool>,
    pub can_restrict_members: Option<bool>,
    pub can_promote_members: Option<bool>,
    pub can_change_info: Option<bool>,
    pub can_invite_users: Option<bool>,
    pub can_post_messages: Option<bool>,
    pub can_edit_messages: Option<bool>,
    pub can_pin_messages: Option<bool>,
    pub can_manage_topics: Option<bool>,
    pub can_post_stories: Option<bool>,
    pub can_edit_stories: Option<bool>,
    pub can_delete_stories: Option<bool>,
    pub can_manage_direct_messages: Option<bool>,
    pub can_manage_tags: Option<bool>,
    /* private fields */

}

Parameters for a promoteChatMember request.

Fields
is_anonymous: Option<bool>

true to make the admin anonymous.

can_manage_chat: Option<bool>

Allows the admin to manage the chat.

can_delete_messages: Option<bool>

Allows the admin to delete messages of other users.

can_manage_video_chats: Option<bool>

Allows the admin to manage video chats.

can_restrict_members: Option<bool>

Allows the admin to restrict, ban, or unban chat members.

can_promote_members: Option<bool>

Allows the admin to add new administrators.

can_change_info: Option<bool>

Allows the admin to change the chat title, photo, and other settings.

can_invite_users: Option<bool>

Allows the admin to invite new users to the chat.

can_post_messages: Option<bool>

Allows the admin to post messages in channels.

can_edit_messages: Option<bool>

Allows the admin to edit messages in channels.

can_pin_messages: Option<bool>

Allows the admin to pin messages.

can_manage_topics: Option<bool>

Allows the admin to manage forum topics.

can_post_stories: Option<bool>

Allows the admin to post stories.

can_edit_stories: Option<bool>

Allows the admin to edit stories posted by others.

can_delete_stories: Option<bool>

Allows the admin to delete stories.

can_manage_direct_messages: Option<bool>

Allows the admin to manage direct messages.

can_manage_tags: Option<bool>

Allows the admin to manage tags.