rustigram_api::methods

Module sending

Message and media sending methods.

Structs

AnswerCallbackQuery

Builder for the answerCallbackQuery method.

CopyMessage

Builder for the copyMessage method.

CopyMessages

Builder for the copyMessages method.

DeleteEphemeralMessage

Builder for the deleteEphemeralMessage method.

DeleteMessage

Builder for the deleteMessage method.

DeleteMessages

Builder for the deleteMessages method.

ForwardMessage

Builder for the forwardMessage method.

ForwardMessages

Builder for the forwardMessages method.

MediaSendOptions

Common optional parameters shared by most media-send methods.

SendAnimation

Builder for the sendAnimation method.

SendAudio

Builder for the sendAudio method.

SendChatAction

Builder for the sendChatAction method.

SendChecklist

Builder for the sendChecklist method.

SendContact

Builder for the sendContact method.

SendDice

Builder for the sendDice method.

SendDocument

Builder for the sendDocument method.

SendGame

Builder for the sendGame method.

SendLivePhoto

Builder for the sendLivePhoto method.

SendLocation

Builder for the sendLocation method.

SendMediaGroup

Builder for the sendMediaGroup method.

SendMessage

Builder for the sendMessage method.

SendMessageDraft

Builder for the sendMessageDraft method. Streams a partial message to the user while it is being generated (Bot API 9.5+).

SendPaidMedia

Builder for the sendPaidMedia method.

SendPhoto

Builder for the sendPhoto method.

SendPoll

Builder for the sendPoll method.

SendRichMessage

Builder for the sendRichMessage method.

SendRichMessageDraft

Builder for the sendRichMessageDraft method.

SendSticker

Builder for the sendSticker method.

SendVenue

Builder for the sendVenue method.

SendVideo

Builder for the sendVideo method.

SendVideoNote

Builder for the sendVideoNote method.

SendVoice

Builder for the sendVoice method.

StopPoll

Builder for the stopPoll method.

Enums

ChatAction

The chat action to display while the bot is preparing a response.